Initial commit
This is a buildroot customization project based on the indications
described in the (*) document.
This project allows to keep package recipes, board support and
configuration files outside of the Buildroot tree, while still
having them nicely integrated in the build logic.
I will start with customizations for the beaglebone board and I
hope to continue over time with the addition of customizations
for other boards.

(*) https://buildroot.org/downloads/manual/customize-outside-br.txt

Signed-off-by: Dario Binacchi <dariobin@libero.it>